feat: Defi Mint & Redeem UI (yRUNE & yTCY) (#3331)
* feat: added Mint UI

* feat: added Redeem UI

* feat: added slippage

+++ VultisigApp/VultisigApp/Features/FunctionTransaction/Bond/BondTransactionScreen.swift
@@ -17,7 +17,7 @@ struct BondTransactionScreen: View {
     
     @State var focusedFieldBinding: FocusedField? = .none
     @FocusState private var focusedField: FocusedField?
-    @State var percentageSelected: Int?
+    @State var percentageSelected: Double?
     
     var body: some View {
         TransactionFormScreen(
