+++ VultisigApp/VultisigApp/Features/FunctionTransaction/Bond/BondTransactionViewModel.swift
@@ -81,7 +81,7 @@ final class BondTransactionViewModel: ObservableObject, Form {
         )
     }
     
-    func onPercentage(_ percentage: Int) {
+    func onPercentage(_ percentage: Double) {
         isMaxAmount = percentage == 100
     }
 }
