+++ VultisigApp/VultisigApp/Features/FunctionTransaction/Unbond/UnbondTransactionViewModel.swift
@@ -62,7 +62,7 @@ final class UnbondTransactionViewModel: ObservableObject, Form {
         )
     }
     
-    func onPercentage(_ percentage: Int) {
+    func onPercentage(_ percentage: Double) {
         isMaxAmount = percentage == 100
     }
